+++ streaming-service/src/main/java/dev/studentpp1/streamingservice/movies/dto/MovieDto.java
@@ -1,10 +1,12 @@
 package dev.studentpp1.streamingservice.movies.dto;
 
+import java.math.BigDecimal;
+
 public record MovieDto(
         Long id,
         String title,
         String description,
         Integer year,
-        Double rating,
+        BigDecimal rating,
         Long directorId
 ) {}
