+++ streaming-service/src/main/java/dev/studentpp1/streamingservice/movies/dto/MovieRequest.java
@@ -2,6 +2,8 @@ package dev.studentpp1.streamingservice.movies.dto;
 
 import jakarta.validation.constraints.*;
 
+import java.math.BigDecimal;
+
 public record MovieRequest(
         @NotBlank(message = "Title is required")
         @Size(max = 255, message = "Title must be less than 255 characters")
@@ -15,7 +17,7 @@ public record MovieRequest(
 
         @DecimalMin(value = "0.0", message = "Rating must be at least 0.0")
         @DecimalMax(value = "10.0", message = "Rating must be at most 10.0")
-        Double rating,
+        BigDecimal rating,
 
         @NotNull(message = "Director ID is required")
         Long directorId
